requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
email_sender==1.2
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2024.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
mailgun==0.1.1
Mako==1.3.5
MarkupSafe==2.1.5
oauthlib==3.2.2
openai==1.50.2
packaging==24.1
pyactiveresource==1.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyhunter==1.7
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tldextract==5.1.2
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
